chore: isolate storage custom utilities
