+++ bd_api/apps/api/v1/validators.py
@@ -5,7 +5,6 @@ import re
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
-from django.core.validators import FileExtensionValidator
 
 current_site = os.getenv("url")
 
@@ -33,11 +32,3 @@ def validate_is_valid_area_key(
             f"Area '{value}' is not valid. Area must valid.\
             See {current_site}/schemas/bd_spatial_coverage_tree/ for valid areas."
         )
-
-
-def validate_is_valid_image_format(value):
-    """Validate image file."""
-    validate_image_file_extension = FileExtensionValidator(
-        allowed_extensions=["jpg", "jpeg", "png", "gif"]
-    )
-    validate_image_file_extension(value)
